refactor: move decision_logs to synthesis/decision_logs.rs
